move component to folder relative
