+++ frontend/src/components/ResultsView/ItemMenu/ItemMenu.tsx
@@ -13,7 +13,7 @@ import { dismissToast, errorToast } from "../../shared/toasts/toasts";
 import { validString } from "../../../utilities/sanitize";
 import { ERROR_INPUT_EVENT } from "../../../utilities/variables";
 import TotalsMenu from "./TotalsMenu/TotalsMenu";
-import EventActions from "./PersonCard/EventActions/EventActions";
+import EventActions from "./EventActions/EventActions";
 
 function ItemMenu() {
   const didMountOnce = didMount();
